+++ Contents/mods/PhunMart/media/lua/client/PhunMart_CShopGlobalObject.lua
@@ -1,17 +1,17 @@
+if isServer() then return end
+
 require "Map/CGlobalObject"
 
 CPhunMartObject = CGlobalObject:derive("CPhunMartObject")
 local inc = 0
 function CPhunMartObject:new(luaSystem, globalObject)
     inc = inc + 1
-    print("+++++++++++CPhunMartObject:new ", tostring(inc))
     local o = CGlobalObject.new(self, luaSystem, globalObject)
 
     return o
 end
 
 function CPhunMartObject:getObject()
-    print("CPhunMartObject:getObject")
     return self:getIsoObject()
 end
